+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
 ## Unreleased
-- Support `readOnly: true` in request body validation
+- Properties with `readOnly: true` are remove from request bodies during request validation.
+- An error is raised if response body includes a property with `writeOnly: true`.
 
 ## 0.13.2
 - Return indicator (`source: { parameter: 'list/1' }`) in error response body when array item in query parameter is invalid